requirements.txt
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
binance-connector==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==39.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pkg_resources==0.0.0
pprintpp==0.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.0.0
python-dotenv==1.0.0
requests==2.28.2
service-identity==21.1.0
six==1.16.0
style==1.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
update==0.0.1
urllib3==1.26.14
zope.interface==5.5.2
